Update ox-wintersmith to not use TODO states for defining published vs. drafts
I use TODO states for tasks that can be archived.

So for blog posts, I decided to use a combination of tags and properties.

Properties are the authoritative data point that determines whether or not a 
post should be published and exported (via export-all). The tags just reflect 
the property state, and are updated whenever an export-all or export-to-blog 
command is invoked.